first version
- use simple scrollview for dashboard
- show separate row for courses
