+++ iOS/ViewControllers/Dashboard/CourseDatesListViewController.swift
@@ -10,10 +10,6 @@ import UIKit
 
 class CourseDatesListViewController: UITableViewController {
 
-    @IBOutlet private var loginButton: UIBarButtonItem!
-
-    var courseActivityViewController: CourseActivityViewController?
-
     var resultsController: NSFetchedResultsController<CourseDate>!
     var resultsControllerDelegateImplementation: TableViewResultsControllerDelegateImplementation<CourseDate>!
 
@@ -24,14 +20,6 @@ class CourseDatesListViewController: UITableViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-
-        // register custom section header view
-        self.tableView.register(R.nib.courseDateHeader(), forHeaderFooterViewReuseIdentifier: R.nib.courseDateHeader.name)
-
-        // setup pull to refresh
-        let refreshControl = UIRefreshControl()
-        refreshControl.addTarget(self, action: #selector(refresh), for: .valueChanged)
-        self.tableView.refreshControl = refreshControl
 
         // setup table view data
         let reuseIdentifier = R.reuseIdentifier.courseDateCell.identifier
@@ -56,47 +44,11 @@ class CourseDatesListViewController: UITableViewController {
         self.setupEmptyState()
     }
 
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        TrackingHelper.shared.createEvent(.visitedDashboard)
-    }
-
     func setupEmptyState() {
         tableView.emptyDataSetSource = self
         tableView.emptyDataSetDelegate = self
         tableView.tableFooterView = UIView()
         tableView.reloadEmptyDataSet()
-    }
-
-    @objc func refresh() {
-        self.tableView.reloadEmptyDataSet()
-        let deadline = UIRefreshControl.minimumSpinningTime.fromNow
-        let stopRefreshControl = {
-            DispatchQueue.main.asyncAfter(deadline: deadline) {
-                self.tableView.refreshControl?.endRefreshing()
-            }
-        }
-
-        CourseHelper.syncAllCourses().onComplete { _ in
-            CourseDateHelper.syncAllCourseDates().onComplete { _ in
-                stopRefreshControl()
-            }
-        }
-
-    }
-
-    override func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        return tableView.dequeueReusableHeaderFooterView(withIdentifier: R.nib.courseDateHeader.name)
-    }
-
-    override func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        return 50
-    }
-
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if let typedInfo = R.segue.courseDatesListViewController.embedCourseActivity(segue: segue) {
-            self.courseActivityViewController = typedInfo.destination
-        }
     }
 
 }
@@ -149,10 +101,6 @@ extension CourseDatesListViewController: DZNEmptyDataSetSource, DZNEmptyDataSetD
         let description = NSLocalizedString("empty-view.course-dates.no-dates.description",
                                             comment: "description for empty course dates list if logged in")
         return NSAttributedString(string: description)
-    }
-
-    func emptyDataSet(_ scrollView: UIScrollView!, didTap view: UIView!) {
-        self.refresh()
     }
 
 }
